chore(rule): move components to rule dir
