+++ ui/src/alerting/components/AlertingIndex.tsx
@@ -6,7 +6,7 @@ import {connect} from 'react-redux'
 import {Grid, GridRow, GridColumn, Page} from '@influxdata/clockface'
 import PageTitleWithOrg from 'src/shared/components/PageTitleWithOrg'
 import ChecksColumn from 'src/alerting/components/ChecksColumn'
-import RulesColumn from 'src/alerting/components/notifications/RulesColumn'
+import RulesColumn from 'src/notifications/rules/components/RulesColumn'
 import EndpointsColumn from 'src/alerting/components/EndpointsColumn'
 import GetAssetLimits from 'src/cloud/components/GetAssetLimits'
 import AssetLimitAlert from 'src/cloud/components/AssetLimitAlert'
