+++ ui/src/index.tsx
@@ -85,8 +85,8 @@ import DEDeleteDataOverlay from 'src/dataExplorer/components/DeleteDataOverlay'
 import NewThresholdCheckEO from 'src/alerting/components/NewThresholdCheckEO'
 import NewDeadmanCheckEO from 'src/alerting/components/NewDeadmanCheckEO'
 import EditCheckEO from 'src/alerting/components/EditCheckEO'
-import NewRuleOverlay from 'src/alerting/components/notifications/NewRuleOverlay'
-import EditRuleOverlay from 'src/alerting/components/notifications/EditRuleOverlay'
+import NewRuleOverlay from 'src/notifications/rules/components/NewRuleOverlay'
+import EditRuleOverlay from 'src/notifications/rules/components/EditRuleOverlay'
 import NewEndpointOverlay from 'src/notifications/endpoints/components/NewEndpointOverlay'
 import EditEndpointOverlay from 'src/notifications/endpoints/components/EditEndpointOverlay'
 import NoOrgsPage from 'src/organizations/containers/NoOrgsPage'
