+++ ui/src/notifications/rules/components/NewRuleOverlay.scss
