+++ ui/src/notifications/rules/components/RuleMessageContents.tsx
