+++ ui/src/notifications/rules/components/RuleOverlayFooter.scss
