+++ ui/src/notifications/rules/components/SMTPMessage.tsx
