+++ ui/src/style/chronograf.scss
@@ -115,7 +115,7 @@
 @import 'src/shared/components/EventMarkerTooltip.scss';
 @import 'src/shared/components/DeleteDataForm/DeleteDataForm.scss';
 @import 'src/shared/components/cloud/CloudOnly.scss';
-@import 'src/alerting/components/notifications/NewRuleOverlay.scss';
+@import 'src/notifications/rules/components/NewRuleOverlay.scss';
 @import 'src/shared/components/dashed_button/DashedButton.scss';
 @import 'src/alerting/components/AlertingIndex.scss';
 @import 'src/alerting/components/CheckHistory.scss';
@@ -125,7 +125,7 @@
 @import 'src/alerting/components/AlertHistoryControls.scss';
 @import 'src/alerting/components/LevelTableField.scss';
 @import 'src/alerting/components/SentTableField.scss';
-@import 'src/alerting/components/notifications/RuleOverlayFooter.scss';
+@import 'src/notifications/rules/components/RuleOverlayFooter.scss';
 @import 'src/alerting/components/SearchBar.scss';
 @import 'src/clientLibraries/components/ClientLibraryOverlay.scss';
 @import 'src/dashboards/components/DashboardsCardGrid.scss';
